requirements.txt
flask==2.0.1
werkzeug==2.0.1
mysql-connector-python==8.0.26
waitress==2.0.0
gunicorn==20.1.0
